app-with-bigquery/flask-bigquery-app/requirements.txt
Flask==2.2.3
google-cloud-bigquery==2.34.0
flask-wtf==1.0.0
wtforms==3.0.1